exports Collection and Document direclty
