+++ src/firestore/context.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const FirestoreContext = React.createContext();
+
+export default FirestoreContext;
